+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## [0.2.0]
+
+### Changes
+
+* Flip part image before applying rotation
+* Some optimizations
+* Removed the [-e] option to include empty frames
+* Removed the [-d] divider option
+* Promisification of callbacks
+
 ## [0.1.0]
 
 ### Notes